@@ -1,78 +1,46 @@
 {
-  "name": "@teleoscope/teleoscope",
-  "version": "0.0.12",
+  "name": "frontend",
+  "version": "1.0.0",
+  "description": "[MongoDB](https://www.mongodb.com/) is a general purpose, document-based, distributed database built for modern application developers and for the cloud era. This example will show you how to connect to and use MongoDB as your backend for your Next.js app.",
+  "main": "babel.config.js",
+  "directories": {
+    "test": "tests"
+  },
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "test": "jest",
     "start": "next start"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Teleoscope/Teleoscope.git"
-  },
-  "publishConfig": {
-    "@teleoscope:registry": "https://npm.pkg.github.com"
-  },
+  "author": "",
+  "license": "ISC",
   "dependencies": {
-    "@cloudamqp/amqp-client": "^1.3.2",
-    "@emotion/react": "^11.8.1",
-    "@emotion/styled": "^11.8.1",
-    "@fontsource/roboto": "^4.5.0",
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
-    "@material-ui/lab": "^4.0.0-alpha.57",
-    "@mui/icons-material": "^5.4.4",
-    "@mui/material": "^5.4.4",
-    "@mui/styled-engine-sc": "^5.4.2",
-    "@reduxjs/toolkit": "^1.8.2",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
+    "@fontsource/roboto": "^4.5.7",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/material": "^5.8.6",
+    "@mui/styles": "^5.8.6",
+    "@reduxjs/toolkit": "^1.8.3",
     "@stomp/stompjs": "^6.1.2",
-    "bcryptjs": "^2.4.3",
-    "celery-node": "^0.5.8",
-    "crypto-js": "^4.1.1",
-    "crypto-random-string": "^4.0.0",
     "draft-js": "^0.11.7",
-    "fs": "^0.0.1-security",
     "lodash": "^4.17.21",
-    "mongodb": "^3.7.3",
-    "net": "^1.0.2",
-    "next": "^12.1.6",
-    "prompt": "^1.3.0",
-    "randomstring": "^1.2.2",
+    "mongodb": "^4.7.0",
+    "next": "^12.2.0",
+    "react": "^18.2.0",
     "react-cookie": "^4.1.1",
     "react-cool-dimensions": "^2.0.7",
-    "react-dnd": "^15.1.1",
-    "react-dnd-html5-backend": "^15.1.2",
-    "react-draggable": "^4.4.3",
+    "react-dom": "^18.2.0",
     "react-grid-layout": "^1.3.4",
-    "react-redux": "^7.2.8",
-    "react-selectable-fast": "^3.4.0",
+    "react-redux": "^8.0.2",
     "redux": "^4.2.0",
     "redux-thunk": "^2.4.1",
-    "swr": "^0.5.5",
-    "websocket": "^1.0.34",
-    "ws": "^8.6.0"
+    "swr": "^1.3.0",
+    "websocket": "^1.0.34"
   },
-  "license": "MIT",
   "devDependencies": {
-    "2": "^3.0.0",
-    "@babel/core": "^7.18.2",
-    "@babel/preset-env": "^7.18.2",
     "@types/bson": "^4.2.0",
-    "babel-jest": "^28.1.0",
-    "babel-plugin-styled-components": "^1.12.0",
-    "install": "^0.13.0",
-    "jest": "^28.1.0",
-    "npm": "^8.12.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "styled-components": "^5.3.3",
-    "ts-jest": "^28.0.4"
-  },
-  "bundledDependencies": [
-    "next"
-  ],
-  "bundleDependencies": [
-    "next"
-  ]
+    "jest": "^28.1.2",
+    "react-test-renderer": "^18.2.0"
+  }
 }
