@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 
 // MUI imports
-import TextField from "@material-ui/core/TextField";
+import TextField from "@mui/material/TextField";
 
 // actions
 import { useSelector, useDispatch } from "react-redux";
