@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 
 // MUI imports
-import ListItemText from "@material-ui/core/ListItemText";
+import ListItemText from "@mui/material/ListItemText";
 import Typography from '@mui/material/Typography';
 import Tooltip from '@mui/material/Tooltip';
 
