@@ -1,7 +1,9 @@
-import { connectToDatabase } from "../../../util/mongodb";
+import clientPromise from '../../../util/mongodb';
 
 export default async (req, res) => {
-  const { db } = await connectToDatabase();
+  const client = await clientPromise;
+  const db = await client.db('aita');
+  
   const { cleanposts } = req.query;
   var queries = [];
   if (!cleanposts) {
