@@ -1,7 +1,8 @@
-import { connectToDatabase } from "../../../util/mongodb";
+import clientPromise from '../../../util/mongodb';
 
 export default async (req, res) => {
-  const { db } = await connectToDatabase();
+  const client = await clientPromise;
+  const db = await client.db('aita');
   const { user } = req.query;
   const user_data = await db.collection("users").findOne({username: user});
   res.json(user_data);
