@@ -1,7 +1,8 @@
-import { connectToDatabase } from "../../../util/mongodb";
+import clientPromise from '../../../util/mongodb';
 
 export default async (req, res) => {
-  const { db } = await connectToDatabase();
+  const client = await clientPromise;
+  const db = await client.db('aita');
   const users = await db.collection("users").find({}).limit(20).toArray();
   res.json(users);
 };
