@@ -3,7 +3,6 @@ import Head from "next/head";
 
 import { Provider } from "react-redux";
 import { CookiesProvider } from "react-cookie";
-import { connectToDatabase } from "../util/mongodb";
 import { SWRConfig } from 'swr'
 
 // store
@@ -11,6 +10,8 @@ import store from "../stores/store";
 
 // custom components
 import Workspace from "../components/Workspace";
+
+import clientPromise from '../util/mongodb';
 
 // contexts
 import { StompContext, client } from "../context/StompContext"
@@ -47,11 +48,37 @@ export default function Home({ isConnected }) {
   );
 }
 
-// Connect to MongoDB
+
+
 export async function getServerSideProps(context) {
-  const { client } = await connectToDatabase();
-  const isConnected = await client.isConnected();
-  return {
-    props: { isConnected },
-  };
+  try {
+    await clientPromise
+    // `await clientPromise` will use the default database passed in the MONGODB_URI
+    // However you can use another database (e.g. myDatabase) by replacing the `await clientPromise` with the following code:
+    //
+    // `const client = await clientPromise`
+    // `const db = client.db("myDatabase")`
+    //
+    // Then you can execute queries against your database like so:
+    // db.find({}) or any of the MongoDB Node Driver commands
+
+    return {
+      props: { isConnected: true },
+    }
+  } catch (e) {
+    console.error(e)
+    return {
+      props: { isConnected: false },
+    }
+  }
 }
+// 
+// 
+// Connect to MongoDB
+// export async function getServerSideProps(context) {
+//   const { client } = await connectToDatabase();
+//   const isConnected = await client.isConnected();
+//   return {
+//     props: { isConnected },
+//   };
+// }
